+++ FFNN/Base/params.py
@@ -1,0 +1,14 @@
+# Here are the means for the Gaussians.
+# These are used when making the random potential
+means_cell: list[list[float, float]] = [
+    [0.74507886, 0.3602802],
+    [0.40147605, 0.06139579],
+    [0.94162198, 0.46722697],
+    [0.79110703, 0.8973808],
+    [0.64732527, 0.07095655],
+    [0.10083943, 0.31935057],
+    [0.24929806, 0.60499613],
+    [0.11377013, 0.42598647],
+    [0.85163671, 0.26495608],
+    [0.18439795, 0.31438099],
+]
