+++ FFNN/TranssferLearning/OInitial_x_0final_t_1alpha_1width_40epochs_25000grid_size_400Network_state.pth
