+++ FFNN/TranssferLearning/TL/Initial_x_0Initial_y_1_0.55final_t_1alpha_1width_40epochs_25500epochsTL_25000grid_size_400lossTL.p
