+++ FFNN/TranssferLearning/TL/Initial_x_0final_t_1alpha_1width_40epochs_25500epochsTL_25000grid_size_400Network_stateTL.p
