+++ FFNN/TranssferLearning/TL/Initial_x_0final_t_1alpha_1width_40epochs_25500grid_size_400Trajectories_y_TL.p
