+++ FFNN/TranssferLearning/terminal_output.pages
