+++ README.md
@@ -1,2 +1,17 @@
 # Transfer-Learning-with-PINNs-for-Efficient-Simulation-of-Branched-Flows
-Code for the Paper "Transfer Learning with Physics-Informed Neural Networks for Efficient Simulation of Branched Flows", Machine Learning and the Physical Sciences Workshop at the 36th conference on Neural Information Processing Systems (NeurIPS)
+
+Code for the Paper "Transfer Learning with Physics-Informed Neural Networks for Efficient Simulation of Branched Flows", Machine Learning and the Physical Sciences Workshop at the 36th conference on Neural Information Processing Systems (NeurIPS).
+
+How to run:
+
+create a virtual environment:
+
+conda create -n <your_venv> python=3.11
+
+conda activate your_environment
+
+pip install -r requirements
+
+
+
+
